lib/posix-user: Add more group file related functions
This patch introduces the changes needed for iterating through the
"/etc/groups" file. But because we currently provide only a single
hardcoded group, we will return just one single entry. This can be
extended easily, either by adding more hardcoded groups or by using
the actual "/etc/groups". However, the latter option is subject to
future work, when we will need a more consistent analysis regarding
user management and permissions on Unikraft.

The groups iterator is thread safe, therefore we also introduce the
`__uk_tls`, qualifier which depends on `CONFIG_HAVE_SCHED`, and mark
the users iterator as TLS as well. Another minor change revisits this
library constructor by marking it as a Unikraft constructor.

Signed-off-by: Costin Lupu <costin.lupu@cs.pub.ro>
Reviewed-by: Simon Kuenzer <simon.kuenzer@neclab.eu>