posix-socket: Provide generics for driver-level registration
This commit introduces the primitives for a Unikraft driver-based
system that attaches a particular address family (AF) to a suite
of posix-socket functions.  For now, we only indicate an
initialization function which can be used by the library providing
the AF number for setup purposes.

Signed-off-by: Alexander Jung <alexander.jung@neclab.eu>
Checkpatch-Ignore: COMPLEX_MACRO
Checkpatch-Ignore: OPEN_BRACE