posix-socket: Register file descriptors with vfscore
This commit provides a mechanism to register new file descriptors
with vfscore.  This will allow the socket, socketpair, and other
methods part of the POSIX socket family which return a file
descriptor to do so correctly and within the Unikraft framework.

To accomplish this, the `socket_alloc_fd` (to create) and
`posix_socket_file_get` (to retrieve) methods are implemented.

Signed-off-by: Alexander Jung <alexander.jung@neclab.eu>